ft_strtok and its functions in a saperate file and changes made accordingly to the Makefile
