+++ Makefile
@@ -22,6 +22,7 @@ SRC_FILES = main.c\
 			built_ins.c\
 			built_ins2.c\
 			src/parse/tokenizer.c\
+			src/parse/ft_strtok.c\
 
 # Object files
 OBJ_FILES = $(SRC_FILES:.c=.o)
